requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
dnspython==2.8.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fastapi==0.128.0
fpdf==1.7.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
motor==3.7.1
numpy==2.2.6
openai==2.15.0
openpyxl==3.1.5
pandas==2.3.3
pydantic==2.12.5
pydantic_core==2.41.5
pypdf==5.1.0
pymongo==4.16.0
python-docx==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.40.0
celery[redis]==5.4.0
redis==5.2.1
requests